LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/workers/importscripts_mime_local.any.sharedworker-expected.txt

CONSOLE MESSAGE: NotSupportedError: Shared workers are not supported

Harness Error (FAIL), message = NotSupportedError: Shared workers are not supported
